Apply Pumba for enforcing cluster disruptions. And, since there is no need for it anymore, remove the "alternative" network again.

+++ get-logs.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-rm *.log
+if [[ -e *.log ]]; then
+  rm *.log
+fi
 docker exec -t openfire-docker-compose_xmpp1_1 cat /usr/local/openfire/logs/all.log > 1.log
 docker exec -t openfire-docker-compose_xmpp2_1 cat /usr/local/openfire/logs/all.log > 2.log
 docker exec -t openfire-docker-compose_xmpp3_1 cat /usr/local/openfire/logs/all.log > 3.log
